Created Common and Basket projects.
