+++ src/Services/Catalog/Catalog.Api/Catalog.Api.csproj
@@ -11,11 +11,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-	<PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.15.0" />
 	<PackageReference Include="MongoDB.Driver" Version="2.15.0" />
-	<PackageReference Include="Swashbuckle.AspNetCore" Version="6.3.0" />
-	<PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="6.3.0" />
-	<PackageReference Include="Hellang.Middleware.ProblemDetails" Version="6.4.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+	<ProjectReference Include="..\..\..\BuildingBlocks\Common.BuildApplication\Common.BuildApplication.csproj" />
   </ItemGroup>
 
 </Project>
