+++ src/YY.eCommerce.sln
@@ -15,6 +15,14 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Basket", "Basket", "{C77F06
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Basket.Api", "Services\Basket\Basket.Api\Basket.Api.csproj", "{BD152B26-6136-435E-A854-9C4F98D0BE92}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Discount", "Discount", "{462FA5E7-9641-42F4-9A1C-D76827D5C25A}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Discount.Api", "Services\Discount\Discount.Api\Discount.Api.csproj", "{B421418F-995C-4E89-B54E-E6A167E5B7DA}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "BuildingBlocks", "BuildingBlocks", "{4E77310E-ABE0-467F-96E8-016BD1656321}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common.BuildApplication", "BuildingBlocks\Common.BuildApplication\Common.BuildApplication.csproj", "{C6CE5EA0-310B-4CFC-995E-2953BD986658}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,6 +41,14 @@ Global
 		{BD152B26-6136-435E-A854-9C4F98D0BE92}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BD152B26-6136-435E-A854-9C4F98D0BE92}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BD152B26-6136-435E-A854-9C4F98D0BE92}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B421418F-995C-4E89-B54E-E6A167E5B7DA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B421418F-995C-4E89-B54E-E6A167E5B7DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B421418F-995C-4E89-B54E-E6A167E5B7DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B421418F-995C-4E89-B54E-E6A167E5B7DA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C6CE5EA0-310B-4CFC-995E-2953BD986658}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C6CE5EA0-310B-4CFC-995E-2953BD986658}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C6CE5EA0-310B-4CFC-995E-2953BD986658}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C6CE5EA0-310B-4CFC-995E-2953BD986658}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -42,6 +58,9 @@ Global
 		{F1CB6913-A06A-4A27-B563-7B756878139B} = {9DC16C6A-09C8-4485-B7DB-E642DE457DE0}
 		{C77F0670-BDB8-4351-B0C0-2CD80C3CBE24} = {782DD9D0-B6F1-4648-9DF2-D890A0DEBEFB}
 		{BD152B26-6136-435E-A854-9C4F98D0BE92} = {C77F0670-BDB8-4351-B0C0-2CD80C3CBE24}
+		{462FA5E7-9641-42F4-9A1C-D76827D5C25A} = {782DD9D0-B6F1-4648-9DF2-D890A0DEBEFB}
+		{B421418F-995C-4E89-B54E-E6A167E5B7DA} = {462FA5E7-9641-42F4-9A1C-D76827D5C25A}
+		{C6CE5EA0-310B-4CFC-995E-2953BD986658} = {4E77310E-ABE0-467F-96E8-016BD1656321}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {0535BE5C-FF1F-497B-900E-1D85C4CF5019}
